+++ src/Mapbender/CoreBundle/Component/Presenter/Application/ConfigService.php
@@ -136,7 +136,9 @@ class ConfigService
             if (!$configGenerator->isInstanceEnabled($assignment->getInstance())) {
                 continue;
             }
-            $configs[] = $configGenerator->getConfiguration($layerset->getApplication(), $assignment->getInstance());
+            $idPrefix = $assignment instanceof ReusableSourceInstanceAssignment ? $assignment->getId().'_' : null;
+            $configuration = $configGenerator->getConfiguration($layerset->getApplication(), $assignment->getInstance(), $idPrefix);
+            $configs[] = $configuration;
         }
         return $configs;
     }