+++ src/Mapbender/WmtsBundle/Component/Presenter/ConfigGeneratorCommon.php
@@ -27,10 +27,10 @@ abstract class ConfigGeneratorCommon extends SourceInstanceConfigGenerator
     abstract protected function getLayerTreeOptions(SourceInstanceItem $instanceLayer);
 
 
-    public function getConfiguration(Application $application, SourceInstance $sourceInstance): array
+    public function getConfiguration(Application $application, SourceInstance $sourceInstance, ?string $idPrefix = null): array
     {
         /** @var WmtsInstance $sourceInstance */
-        return array_merge(parent::getConfiguration($application, $sourceInstance), [
+        return array_merge(parent::getConfiguration($application, $sourceInstance, $idPrefix), [
             'version' => $sourceInstance->getSource()->getVersion(),
             'options' => $this->getOptionsConfiguration($sourceInstance),
             'children' => $this->getRootLayerConfig($sourceInstance),